Refactoring
- extracted sections in own files
- added more languages
- changed group picture
