+++ pages/_app.tsx
@@ -1,10 +1,17 @@
-import type { AppProps } from 'next/app'
-import { appWithTranslation } from 'next-i18next'
+import type { AppProps } from 'next/app';
+import { appWithTranslation } from 'next-i18next';
+import Head from 'next/head';
 // import nextI18NextConfig from '../next-i18next.config.js'
 
 const MyApp = ({ Component, pageProps }: AppProps) => (
-  <Component {...pageProps} />
-)
+  <>
+    <Head>
+      <title>Taekwon Do Chon-Ji </title>
+      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    </Head>
+    <Component {...pageProps} />
+  </>
+);
 
 // https://github.com/i18next/next-i18next#unserializable-configs
-export default appWithTranslation(MyApp /*, nextI18NextConfig */)
+export default appWithTranslation(MyApp /*, nextI18NextConfig */);
