+++ pages/_document.tsx
@@ -1,6 +1,6 @@
-import Document, { Html, Head, Main, NextScript } from "next/document";
-import type { DocumentProps } from "next/document";
-import i18nextConfig from "../next-i18next.config";
+import Document, { Html, Head, Main, NextScript } from 'next/document';
+import type { DocumentProps } from 'next/document';
+import i18nextConfig from '../next-i18next.config';
 
 type Props = DocumentProps & {
   // add custom document props
@@ -15,19 +15,11 @@ class MyDocument extends Document<Props> {
         <Head>
           <meta charSet="UTF-8" />
           <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-          <meta
-            name="viewport"
-            content="width=device-width, initial-scale=1.0"
-          />
           <link rel="stylesheet" href="app.css" />
           <link
             rel="stylesheet"
             href="https://unicons.iconscout.com/release/v4.0.0/css/line.css"
           />
-        
-         
-
-          <title>Document</title>
         </Head>
         <body>
           <Main />
