+++ public/locales/it/common.json
@@ -12,7 +12,7 @@
     "about-us": "About us",
     "services": "Services",
     "trainers": "Trainers",
-    "contact-us": "Contact us en"
+    "contact-us": "Contact us es"
   },
   "home": {
     "title": "WELCOME TO chon-ji school",
